+++ src/test/java/de/qaware/heimdall/config/ConfigCoderImplTest.java
@@ -26,7 +26,7 @@ package de.qaware.heimdall.config;
 import org.testng.annotations.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.collection.IsArrayContaining.hasItemInArray;
+import static org.hamcrest.Matchers.hasItemInArray;
 import static org.hamcrest.core.Is.is;
 
 /**