+++ src/test/java/de/qaware/heimdall/util/Base64Test.java
@@ -3,12 +3,13 @@ package de.qaware.heimdall.util;
 import org.testng.annotations.Test;
 
 import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 
 public class Base64Test {
-    private static final Charset UTF_8 = Charset.forName("UTF-8");
+    private static final Charset UTF_8 = StandardCharsets.UTF_8;
 
     @Test
     public void testDecode() throws Exception {